Remove the second line edit asking for a password
I'm not sure the reasoning behind having two line edits for the wifi
passwords. As far as I can tell, it only adds frustration:

1) Multiple areas where the password can be wrong (on the two line edits
AND when connecting)
2) The show checkbox obviates the need for making sure it's entered the
same twice.

At the same time, I've alse run into the issue where I get this dialog
and want to quit out of it, but I can't. So, I've redesigned them to be
a more traditional OK and Cancel button. This allows us to exit out of
the dialog if we really don't want to do any changes.